+++ Libplanet.Tests/Store/LiteDBStoreTest.cs
@@ -1,13 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Security.Cryptography;
-using Libplanet.Action;
-using Libplanet.Blocks;
-using Libplanet.Tests.Common.Action;
-using Libplanet.Tx;
-using Org.BouncyCastle.Crypto;
-using Xunit;
 
 namespace Libplanet.Tests.Store
 {